Refactor: API 호출 일관성 있게 수정
